feat: [#188121854] generate dynamic site map, includes starter kits
