+++ web/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "airtable:dump": "node src/scripts/airtableDump.js",
     "airtable:fetch-opps": "node src/scripts/airtableOpportunityDump.js",
-    "build": "yarn next build",
+    "build": "yarn next build && yarn run sitemap",
     "build-storybook": "storybook build",
     "clean": "rimraf .next public/vendor",
     "cypress:open": "cypress open",
@@ -22,6 +22,7 @@
     "prepare:deploy": "next export",
     "prettier": "prettier --write . --ignore-path=./.eslintignore",
     "prettier:check": "prettier . --check --ignore-path=./.eslintignore",
+    "sitemap": "yarn dlx tsx src/scripts/generateSitemap.ts",
     "snapshot:update": "jest --updateSnapshot",
     "spellcheck": "cspell '**/*.md' '**/*.yml' --config ../cspell.json",
     "start": "next start",
