+++ web/src/lib/utils/starterKits.ts
@@ -1,0 +1,27 @@
+import { Industries } from "@businessnjgovnavigator/shared/industry";
+
+export const STARTER_KITS_GENERIC_SLUG = "nj-business";
+
+type PathParameters<P> = { params: P; locale?: string };
+
+export type StarterKitsUrl = {
+  starterKitsUrl: string;
+};
+
+export const getAllStarterKitUrls = (): PathParameters<StarterKitsUrl>[] => {
+  return Industries.filter((industry) => industry.isEnabled).map((industry) => {
+    if (industry.id === "generic") {
+      return {
+        params: {
+          starterKitsUrl: STARTER_KITS_GENERIC_SLUG,
+        },
+      };
+    }
+
+    return {
+      params: {
+        starterKitsUrl: industry.id,
+      },
+    };
+  });
+};
