+++ web/tsconfig.json
@@ -29,6 +29,6 @@
     "types": ["@types/testing-library__jest-dom", "@types/jest", "jest", "node"],
     "incremental": true
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "../web/src/lib/utils/starterKits.ts"],
   "exclude": ["node_modules", "cypress"]
 }
